refactor: extract system routes (health, log-level, disk-stats, diagnostics, update-check) into routes/system.ts
Pure structural move of the system/operational domain out of createApp; no
handler logic changed. Moved walkCaseFiles and the update-check/preflight
domain-local helpers alongside their routes.

Graduated onto RouteContext (each also consumed by routes staying in server.ts):
- appStartedAt, recentImportFailures, recentAiErrors, hasAiProvider() — the
  diagnostics rings/recorders shared with the import + AI error-recording paths.
- captureBuffers(), synthInFlight() — capture/synthesis state read by /diagnostics.
- importerRegistry(), irisClient() — exposed as live accessors because both are
  reassigned at runtime (importerRegistry after async load; irisClient on
  /iris/reconnect), so a value copy would go stale.
- dropWatchEnabled(), enrichmentProviders(), enrichHealth() — drop/enrichment
  state also used by their own domains; accessors because they are constructed
  after the ctx literal.
